Basic app created and functioning.
I added in a draggable top bar and a close button in case I want to run
"frame: false" in main.js.

Next step is to run a concept test for file saving, then I should be good
to go on creating the whole html interface. After that should be concept
testing for filtering logic and quest box design organization on the
page.