Finished filtering logic
Finished implementing all the filtering logic and integrated it with the
existing html DOM. NOTE: I opted to create a copy of the whole dataset
whenever it's loaded so that I can access it column-wise and row-wise.
Should probably go back and fix that to avoid 2N complexity whenever
loading in datasets.

Next small step is implementing quest card display in concept app. After
that I should be able to start the app's v1.0 design layout.